+++ examples/find_color.py
@@ -3,8 +3,6 @@ from photons_app.special import FoundSerials
 
 from photons_colour import ColourMessages
 
-import binascii
-
 collector = library_setup()
 
 lan_target = collector.configuration['target_register'].resolve("lan")
@@ -13,7 +11,7 @@ async def doit():
     msg = ColourMessages.GetColor()
     async for pkt, _, _ in lan_target.script(msg).run_with(FoundSerials()):
         hsbk = " ".join("{0}={1}".format(key, pkt.payload[key]) for key in ("hue", "saturation", "brightness", "kelvin"))
-        print("{0}: {1}".format(binascii.hexlify(pkt.target[:6]).decode(), hsbk))
+        print("{0}: {1}".format(pkt.serial, hsbk))
 
 loop = collector.configuration["photons_app"].uvloop
 loop.run_until_complete(doit())