+++ docs/start/changelog.rst
@@ -7,4 +7,6 @@ ChangeLog
     Initial opensource release after over a year of internal development.
 
 0.5.1
-    Tile animations
+    * Tile animations
+    * Added a ``serial`` property to packets that returns the hexlified target
+      i.e. "d073d5000001" or None if target isn't set on the packet